+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## not released
+
+### Fixed
+
+ - Default tags were not sent to adapters for metrics declared before `default_tag` declaration. [@Envek]
+
 ## 0.5.0 - 2020-01-29
 
 ### Added